Add Edit Task Description feature.
